Fix: RacingGame 테스트 수정
